Consolidate options
This commit consolidates the contents of the lexer/options.go and
parser/options.go into lexer/lexer.go and parser/parser.go, respectively.
